Display the CAS username on the user profile ("Edit User") form.
If the user has zero or one CAS usernames, we allow the CAS username
to be edited directly on the form. The actual database updating is
done in hook_user_update().

If the user has multiple CAS usernames, we display the CAS usernames
and provide a link to the advanced editing page.