Restore "Initial login redirection" functionality.
Initial login redirection functionality was inadvertently
disabled when the $_SESSION['cas_first_login'] variable
was no longer being set.

A simpler way to achieve the same effect is to just pass
an argument to the cas_login_page() function if the user
just registered.

Tested to redirect on automatic account creation. Does not
redirect, however, if the user was created previously by
an administrator.