Use checkboxes for role selection, to be consistent with core.
The user module uses checkboxes to select user roles.
We implement the same code here for ease and consistency.

The change from using 'select' to 'checkboxes' requires
the addition of array_filter() in the cas_roles()
function, as the variable storage of the selected
items in a checkbox and select list differ.